+++ R/tar_nix_ga.R
@@ -5,6 +5,7 @@
 #' GitHub Actions and will then run the projects \{targets\} pipeline. Make
 #' sure to give read and write permissions to the GitHub Actions bot.
 #' @return Nothing, copies file to a directory.
+#' @family CI/CD
 #' @export
 #' @examples
 #' \dontrun{
@@ -41,6 +42,7 @@ tar_nix_ga <- function() {
     message("GitHub Actions workflow file saved to: ", path)
   }
 
-  if (identical(Sys.getenv("TESTTHAT"), "true"))
+  if (identical(Sys.getenv("TESTTHAT"), "true")) {
     paste0(path, "/run-pipeline.yaml")
+  }
 }